personaforge_backend/requirements.txt
Django==4.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.5
django-cors-headers==4.2.0
django-filter==23.2
psycopg2-binary==2.9.6
python-decouple==3.8
gunicorn
setuptools